python_version/requirements.txt
openai>=1.44.0
python-dotenv>=1.0.1
requests>=2.32.3
mcp>=1.1.0
